# perf: optimize Gvk.ApiVersion()